requirements.txt
flask
python-dotenv
google-generativeai
pymessenger
pandas
openpyxl
gunicorn

